Renames subprojects
