+++ kanonic-syntax-gen/build.gradle.kts
@@ -3,18 +3,18 @@ plugins {
     id("org.gradle.application")
 }
 
-version = "unspecified"
-
 application {
-    applicationName = "kanonic"
-    mainClass.set("io.johnedquinn.kanonic.tool.Main")
+    applicationName = "kanonic-generate"
+    mainClass.set("io.johnedquinn.kanonic.syntax.Main")
 }
 
 distributions {
     main {
-        distributionBaseName.set("kanonic-tool")
+        distributionBaseName.set("kanonic-generate")
     }
 }
+
+version = "unspecified"
 
 repositories {
     mavenCentral()
@@ -24,9 +24,8 @@ dependencies {
     implementation(kotlin("stdlib"))
     implementation(project(":kanonic-gen"))
     implementation(project(":kanonic-runtime"))
-    implementation(project(":kanonic-syntax"))
-    implementation(project(":kanonic-syntax-gen"))
     implementation("info.picocli:picocli:4.7.0")
+    implementation("com.squareup:kotlinpoet:1.12.0")
 }
 
 tasks.named<Tar>("distTar") {
